chore: clean up proposal-list
