+++ package.json
@@ -60,7 +60,7 @@
     "graphql-request": "^7.4.0",
     "humanize-duration": "^3.33.2",
     "jotai": "^2.19.1",
-    "lucide-react": "^0.461.0",
+    "lucide-react": "^1.16.0",
     "mixpanel-browser": "^2.78.0",
     "next-themes": "^0.4.6",
     "react": "catalog:",
